refactor(chat): dedupe system prompt with MCP instructions, split prompt.ts
The chat SYSTEM_PROMPT restated guidance the platform MCP server already
sends via its instructions (appended to the same system prompt): the
run_and_wait exclusivity rules, the validateInlineRun pre-validation ban,
the async launch/wait restatement, and the integration-preference order.
Those paragraphs are deleted; the single source is buildServerInstructions
in the MCP router. Chat keeps only what the server text lacks: persona,
the operation-vs-agent decision tree, the inline manifest example, the
exact run_and_wait input shape, and how a chat reply must consume the
returned result (never fabricate it).

Mechanical split: SYSTEM_PROMPT, CallerContext, formatCallerContext and
buildCallerContextBlock move verbatim to prompt.ts; chat-stream.ts keeps
the stream-step concerns. Also carries the persistAssistantMessage
call-site adaptation (parent chaining) from the pipeline hardening commit.
